feat(BodyPhysics): add body physics script
  > ** Breaking Changes **

The Body Physics script is a replacement for the Player Presence script
and also handles snapping to nearest floor which has now also been
removed from the Height Adjust Teleporter script.

The Height Adjust Teleport script now purely deals with changing the
position of the play area based on the destination marker and does not
attempt to snap to the nearest floor.

The Body Physics script handles creating a rigidbody and collider for
to represent the user and this prevents a user from walking through
items when the play area is being moved around. It also allows for
gravity to make the user fall off objects.

If the user physically walks in the play area then the movement is
detected and the user rigidbody is set to kinematic to allow walking
into objects and prevent the sickly pushing the world away issue.

Also, leaning over objects is allowed by checking to see if the user
is leaning over a valid object and setting the rigidbody to kinematic
so the rigidbody doesn't push the user away.

Snapping to the nearest floor is also handled by the new Body Physics
script and it determines whether to use gravity or to use the
attached teleporter to teleport the user to the new floor.

The fall restrictions based on the controller position is also in the
Body Physics script, so a controller held over the existing floor will
prevent falling to allow peering over ledges.

The relevant example scenes have been updated to include the new
Body Physics script and all other example scenes have been tidied up
a bit so the script order is a bit more uniform.